requirements.txt
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
Django==2.1.4
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
